+++ contrib/opencensus-ext-azure/opencensus/ext/azure/trace_exporter/__init__.py
@@ -15,7 +15,6 @@
 import atexit
 import json
 import logging
-import os
 
 from opencensus.common.schedule import QueueExitEvent
 from opencensus.ext.azure.common import Options, utils
@@ -56,13 +55,14 @@ ERROR_NAME = attributes_helper.COMMON_ATTRIBUTES['ERROR_NAME']
 STACKTRACE = attributes_helper.COMMON_ATTRIBUTES['STACKTRACE']
 
 
-class AzureExporter(BaseExporter, ProcessorMixin, TransportMixin):
+class AzureExporter(BaseExporter, TransportMixin, ProcessorMixin):
     """An exporter that sends traces to Microsoft Azure Monitor.
 
     :param options: Options for the exporter.
     """
 
     def __init__(self, **options):
+        super(AzureExporter, self).__init__(**options)
         self.options = Options(**options)
         utils.validate_instrumentation_key(self.options.instrumentation_key)
         self.storage = None
@@ -75,10 +75,9 @@ class AzureExporter(BaseExporter, ProcessorMixin, TransportMixin):
                 source=self.__class__.__name__,
             )
         self._telemetry_processors = []
-        super(AzureExporter, self).__init__(**options)
         atexit.register(self._stop, self.options.grace_period)
         # start statsbeat on exporter instantiation
-        if not os.environ.get("APPLICATIONINSIGHTS_STATSBEAT_DISABLED_ALL"):
+        if self._check_stats_collection():
             statsbeat.collect_statsbeat_metrics(self.options)
         # For redirects
         self._consecutive_redirects = 0  # To prevent circular redirects